requirements.txt
numpy==1.18.5
pandas==1.2.2
streamlit==0.79.0
scipy==1.6.0
google-api-core==1.26.1
google-auth==1.28.0
google-auth-oauthlib==0.4.2
google-cloud-speech==2.1.0
google-pasta==0.2.0
googleapis-common-protos==1.53.0
